Minor improvements to circle layer example Renamed 'layers/misc' to 'layers/shared'
