+++ lib/flutter_map.dart
@@ -29,15 +29,15 @@ export 'package:flutter_map/src/layer/attribution_layer/rich/widget.dart';
 export 'package:flutter_map/src/layer/attribution_layer/simple.dart';
 export 'package:flutter_map/src/layer/circle_layer/circle_layer.dart';
 export 'package:flutter_map/src/layer/marker_layer/marker_layer.dart';
-export 'package:flutter_map/src/layer/misc/layer_interactivity/layer_hit_notifier.dart';
-export 'package:flutter_map/src/layer/misc/layer_interactivity/layer_hit_result.dart';
-export 'package:flutter_map/src/layer/misc/line_patterns/stroke_pattern.dart';
-export 'package:flutter_map/src/layer/misc/mobile_layer_transformer.dart';
-export 'package:flutter_map/src/layer/misc/translucent_pointer.dart';
 export 'package:flutter_map/src/layer/overlay_image_layer/overlay_image_layer.dart';
 export 'package:flutter_map/src/layer/polygon_layer/polygon_layer.dart';
 export 'package:flutter_map/src/layer/polyline_layer/polyline_layer.dart';
 export 'package:flutter_map/src/layer/scalebar/scalebar.dart';
+export 'package:flutter_map/src/layer/shared/layer_interactivity/layer_hit_notifier.dart';
+export 'package:flutter_map/src/layer/shared/layer_interactivity/layer_hit_result.dart';
+export 'package:flutter_map/src/layer/shared/line_patterns/stroke_pattern.dart';
+export 'package:flutter_map/src/layer/shared/mobile_layer_transformer.dart';
+export 'package:flutter_map/src/layer/shared/translucent_pointer.dart';
 export 'package:flutter_map/src/layer/tile_layer/tile_builder.dart';
 export 'package:flutter_map/src/layer/tile_layer/tile_coordinates.dart';
 export 'package:flutter_map/src/layer/tile_layer/tile_display.dart';
