+++ lib/src/layer/polygon_layer/polygon_layer.dart
@@ -6,8 +6,8 @@ import 'package:dart_earcut/dart_earcut.dart';
 import 'package:flutter/foundation.dart';
 import 'package:flutter/widgets.dart';
 import 'package:flutter_map/flutter_map.dart';
-import 'package:flutter_map/src/layer/misc/layer_interactivity/internal_hit_detectable.dart';
-import 'package:flutter_map/src/layer/misc/line_patterns/pixel_hiker.dart';
+import 'package:flutter_map/src/layer/shared/layer_interactivity/internal_hit_detectable.dart';
+import 'package:flutter_map/src/layer/shared/line_patterns/pixel_hiker.dart';
 import 'package:flutter_map/src/misc/offsets.dart';
 import 'package:flutter_map/src/misc/simplify.dart';
 import 'package:latlong2/latlong.dart' hide Path;
