+++ lib/src/layer/shared/layer_interactivity/internal_hit_detectable.dart
