+++ lib/src/layer/shared/line_patterns/pixel_hiker.dart
