+++ lib/src/layer/shared/line_patterns/stroke_pattern.dart
