+++ lib/src/layer/shared/line_patterns/visible_segment.dart
