+++ test/full_coverage_test.dart
@@ -14,12 +14,12 @@ import 'package:flutter_map/src/layer/attribution_layer/rich/widget.dart';
 import 'package:flutter_map/src/layer/attribution_layer/simple.dart';
 import 'package:flutter_map/src/layer/circle_layer/circle_layer.dart';
 import 'package:flutter_map/src/layer/marker_layer/marker_layer.dart';
-import 'package:flutter_map/src/layer/misc/layer_interactivity/internal_hit_detectable.dart';
-import 'package:flutter_map/src/layer/misc/mobile_layer_transformer.dart';
-import 'package:flutter_map/src/layer/misc/translucent_pointer.dart';
 import 'package:flutter_map/src/layer/overlay_image_layer/overlay_image_layer.dart';
 import 'package:flutter_map/src/layer/polygon_layer/polygon_layer.dart';
 import 'package:flutter_map/src/layer/polyline_layer/polyline_layer.dart';
+import 'package:flutter_map/src/layer/shared/layer_interactivity/internal_hit_detectable.dart';
+import 'package:flutter_map/src/layer/shared/mobile_layer_transformer.dart';
+import 'package:flutter_map/src/layer/shared/translucent_pointer.dart';
 import 'package:flutter_map/src/layer/tile_layer/tile.dart';
 import 'package:flutter_map/src/layer/tile_layer/tile_bounds/tile_bounds.dart';
 import 'package:flutter_map/src/layer/tile_layer/tile_bounds/tile_bounds_at_zoom.dart';
